[Refactor] : domain token 모듈 생성
Git : #23
