+++ eb_fast_api/service/jwt/tests/conftest.py
@@ -1,6 +1,5 @@
 import pytest
 from eb_fast_api.service.jwt.testings.mock_jwt_service import MockJWTService
-from eb_fast_api.service.jwt.sources.jwt_service import JWTService
 
 
 @pytest.fixture(scope="function")
@@ -10,12 +9,3 @@ def mockJWTService():
     yield service
     del service
     print("Del MockJWTService")
-
-
-@pytest.fixture(scope="function")
-def realJWTService():
-    service = JWTService()
-    print("Create RealJWTService")
-    yield service
-    del service
-    print("Del RealJWTService")
